refactor(console): extract component
extract component
